requirements.txt
Flask
python-decouple
safrs
gunicorn
psycopg2